+++ solgen/gen.go
@@ -288,3 +288,36 @@ func main() {
 		fmt.Println("successfully exported abi files")
 	}
 }
+
+func GenerateEspressoTEEContracts(modules map[string]*moduleInfo, parent string) (map[string]*moduleInfo, error) {
+	filePathsEspressoTeeContracts, err := filepath.Glob(filepath.Join(parent, "espresso-tee-contracts", "out", "*.sol", "*.json"))
+	if err != nil {
+		return modules, fmt.Errorf("failed to get path for espresso tee contracts: %w", err)
+	}
+
+	espressoTEEContractsInfo := modules["espressogen"]
+	if espressoTEEContractsInfo == nil {
+		espressoTEEContractsInfo = &moduleInfo{}
+		modules["espressogen"] = espressoTEEContractsInfo
+	}
+
+	for _, path := range filePathsEspressoTeeContracts {
+		_, file := filepath.Split(path)
+		name := file[:len(file)-5]
+
+		data, err := os.ReadFile(path)
+		if err != nil {
+			return modules, fmt.Errorf("could not read %s for contract %s: %w", path, name, err)
+		}
+		artifact := FoundryArtifact{}
+		if err := json.Unmarshal(data, &artifact); err != nil {
+			return modules, fmt.Errorf("failed to parse espresso contract %s: %w", name, err)
+		}
+		espressoTEEContractsInfo.addArtifact(HardHatArtifact{
+			ContractName: name,
+			Abi:          artifact.Abi,
+			Bytecode:     artifact.Bytecode.Object,
+		})
+	}
+	return modules, nil
+}
